Signup: `signupProgressStore` → `signupProgress`
`SignupProgressStore` is a singleton that provides a getter for returning the user's progress in signup. `signupProgressStore` was a prop used to pass the result of `SignupProgressStore.get()` call down to the current step in signup. This name was erroneous, as `signupProgressStore` was not the store itself, but the data it contained. This commit updates every instance of `signupProgressStore` to `signupProgress` in `client/signup`.